Training update, Skill update and modal cancel button redirect fix
Training data can now be updated. Skill data can now be updated. Cancel buttons on all modals on all page now cancel without redirecting to a new page. This issue was resolved by setting type = 'button'. Not setting a type or setting a different type for a button in a form will cause this issue again.